Add feature flag for clustering/federation (#97)
* Add feature flag for clustering/federation

For now, this flag is only allowing or preventing creation of some privileges
in the kiali-controlplane role related to reading some secrets in the
Istio namespace. There is no effect on Kiali server side, because it is
failing gracefully. In case it's needed later, this flag is being passed
to Kiali in its ConfigMap.

Default value is `true` for backwards compatibility.

Related kiali/kiali#4147

Co-authored-by: John Mazzitelli <mazz@redhat.com>

* Feedback: Don't omit creation of role-controlplane