Support for updation of kiali-server helm value view_only_mode
Prior to this change, the post installation update of the kiali-server
helm chart value deployment.view_only_mode fails as it tries to change
the roleRef which is immutable. This change tries to create a new
clusterrolebinding based on the view_only_mode configuration.

Fixes https://github.com/kiali/kiali/issues/6139